rasberrypi: IncrementalEncoder: factor out state machine
