+++ ports/raspberrypi/common-hal/rotaryio/IncrementalEncoder.c
@@ -28,6 +28,7 @@
 
 #include <hardware/regs/pio.h>
 #include "common-hal/rotaryio/IncrementalEncoder.h"
+#include "shared-module/rotaryio/IncrementalEncoder.h"
 #include "bindings/rp2pio/__init__.h"
 #include "bindings/rp2pio/StateMachine.h"
 
@@ -88,12 +89,10 @@ void common_hal_rotaryio_incrementalencoder_construct(rotaryio_incrementalencode
     common_hal_rp2pio_statemachine_run(&self->state_machine, encoder_init, MP_ARRAY_SIZE(encoder_init));
 
     // We're guaranteed by the init code that some output will be available promptly
-    uint8_t state;
-    common_hal_rp2pio_statemachine_readinto(&self->state_machine, &state, 1, 1);
-    // Top two bits of self->last_state don't matter, because they'll be gone as soon as
-    // interrupt handler is called.
-    self->last_state = state & 3;
+    uint8_t quiescent_state;
+    common_hal_rp2pio_statemachine_readinto(&self->state_machine, &quiescent_state, 1, 1);
 
+    shared_module_softencoder_state_init(self, quiescent_state & 3);
     common_hal_rp2pio_statemachine_set_interrupt_handler(&self->state_machine, incrementalencoder_interrupt_handler, self, PIO_IRQ0_INTF_SM0_RXNEMPTY_BITS);
 }
 
@@ -109,67 +108,13 @@ void common_hal_rotaryio_incrementalencoder_deinit(rotaryio_incrementalencoder_o
     common_hal_rp2pio_statemachine_deinit(&self->state_machine);
 }
 
-mp_int_t common_hal_rotaryio_incrementalencoder_get_position(rotaryio_incrementalencoder_obj_t *self) {
-    return self->position;
-}
-
-void common_hal_rotaryio_incrementalencoder_set_position(rotaryio_incrementalencoder_obj_t *self,
-    mp_int_t new_position) {
-    self->position = new_position;
-}
-
 STATIC void incrementalencoder_interrupt_handler(void *self_in) {
     rotaryio_incrementalencoder_obj_t *self = self_in;
-    // This table also works for detent both at 11 and 00
-    // For 11 at detent:
-    // Turning cw: 11->01->00->10->11
-    // Turning ccw: 11->10->00->01->11
-    // For 00 at detent:
-    // Turning cw: 00->10->11->10->00
-    // Turning ccw: 00->01->11->10->00
-
-    // index table by state <oldA><oldB><newA><newB>
-    #define BAD 7
-    static const int8_t transitions[16] = {
-        0,    // 00 -> 00 no movement
-        -1,   // 00 -> 01 3/4 ccw (11 detent) or 1/4 ccw (00 at detent)
-        +1,   // 00 -> 10 3/4 cw or 1/4 cw
-        BAD,  // 00 -> 11 non-Gray-code transition
-        +1,   // 01 -> 00 2/4 or 4/4 cw
-        0,    // 01 -> 01 no movement
-        BAD,  // 01 -> 10 non-Gray-code transition
-        -1,   // 01 -> 11 4/4 or 2/4 ccw
-        -1,   // 10 -> 00 2/4 or 4/4 ccw
-        BAD,  // 10 -> 01 non-Gray-code transition
-        0,    // 10 -> 10 no movement
-        +1,   // 10 -> 11 4/4 or 2/4 cw
-        BAD,  // 11 -> 00 non-Gray-code transition
-        +1,   // 11 -> 01 1/4 or 3/4 cw
-        -1,   // 11 -> 10 1/4 or 3/4 ccw
-        0,    // 11 -> 11 no movement
-    };
 
     while (common_hal_rp2pio_statemachine_get_in_waiting(&self->state_machine)) {
         // Bypass all the logic of StateMachine.c:_transfer, we need something
         // very simple and fast for an interrupt!
-        uint8_t new = self->state_machine.pio->rxf[self->state_machine.state_machine];
-
-        // Shift the old AB bits to the "old" position, and set the new AB bits.
-        self->last_state = (self->last_state & 0x3) << 2 | (new & 0x3);
-
-        int8_t quarter_incr = transitions[self->last_state];
-        if (quarter_incr == BAD) {
-            // Missed a transition. We don't know which way we're going, so do nothing.
-            return;
-        }
-
-        self->quarter_count += quarter_incr;
-        if (self->quarter_count >= 4) {
-            self->position += 1;
-            self->quarter_count = 0;
-        } else if (self->quarter_count <= -4) {
-            self->position -= 1;
-            self->quarter_count = 0;
-        }
+        uint8_t new_state = self->state_machine.pio->rxf[self->state_machine.state_machine];
+        shared_module_softencoder_state_update(self, new_state);
     }
 }
