+++ ports/raspberrypi/mpconfigport.mk
@@ -26,6 +26,7 @@ CIRCUITPY_BITOPS ?= 1
 CIRCUITPY_PWMIO ?= 1
 CIRCUITPY_RGBMATRIX ?= 1
 CIRCUITPY_ROTARYIO ?= 1
+CIRCUITPY_ROTARYIO_SOFTENCODER = 1
 
 # Things that need to be implemented.
 # Use PWM interally
