+++ py/circuitpy_mpconfig.mk
@@ -271,6 +271,9 @@ CFLAGS += -DCIRCUITPY_RGBMATRIX=$(CIRCUITPY_RGBMATRIX)
 CIRCUITPY_ROTARYIO ?= 1
 CFLAGS += -DCIRCUITPY_ROTARYIO=$(CIRCUITPY_ROTARYIO)
 
+CIRCUITPY_ROTARYIO_SOFTENCODER ?= 0
+CFLAGS += -DCIRCUITPY_ROTARYIO_SOFTENCODER=$(CIRCUITPY_ROTARYIO_SOFTENCODER)
+
 CIRCUITPY_RTC ?= 1
 CFLAGS += -DCIRCUITPY_RTC=$(CIRCUITPY_RTC)
